First pass at async listen requests
Fix requirements

Use more asyncio

Add and use check_drop, fix sync

Ensure empty files are created

Don't use DEBUG by default

Travis doesn't like it

Add logging to check_drop

Move scripts

Don't log an error when file is not found

Use asyncio in all the places

fix tests

async cleanup

+++ itests/hostnode.sh
@@ -3,9 +3,17 @@ set -ex -o pipefail
 
 /work/itests/setup.sh
 
-git clone https://github.com/yelp/paasta /small
+if [[ -d "/ext/small" ]]; then
+    cp -r /ext/small /small
+else
+    git clone https://github.com/yelp/paasta /small
+fi
 mkdir /large
-wget -v https://cdimage.debian.org/debian-cd/current/amd64/iso-cd/debian-9.4.0-amd64-netinst.iso -O /large/debian.iso
+if [[ -f "/ext/debian.iso" ]]; then
+    cp /ext/debian.iso /large/
+else
+    wget -v https://cdimage.debian.org/debian-cd/current/amd64/iso-cd/debian-9.4.0-amd64-netinst.iso -O /large/debian.iso
+fi
 
 drop_init /small > /share/small
 drop_init /large > /share/large
